Restructure and test
